Create programReverseWithFor
copy paste for note

Line 6: In the first version of the number counter we reserved a variable for each integer. Later on in the program these variables stored values from user input. On this program, on line six, we do exactly the same thing except this time we reserve an element from the array numbers[] for each value.

Line 7: The program requires a special index variable index, which allows us to manipulate the array elements.

Line 9: Because our program actually repeats the same action a hundred times, we create a loop that repeats itself for a hundred times. The first loop of the program is in charge of asking for user input and storing the input to the array. On line:

for(index=0; index<100; index++)

we state that the for-loop is repeated a hundred times by first assigning the value 0 to the variable index and incrementing that value by one during each iteration until the value is 99. After this the program execution exits the loop.

Line 10: We want the program to print "Input N. number: " on each iteration, where N is the order of the current number. On first iteration we want to print "Input 1. number: ", on second "Input 2. number: " and so on.

Because the value of index is at first zero, we must add one to the value on each iteration to print the correct order number:

cout << "Input " << index+1 << ". number: ";

As you surely notice, when printed, the order is incremented on each iteration.

Line 11: The numbers user inputs are stored in the array numbers[], which size is 100 elements.

When the loop has been executed a hundred times, the first number the user input is stored in the element numbers[0] and the last one in numbers[99].

Line 14: Because we don't want to print "Numbers in reverse order: " a hundred times, the print statement is written outside the previous loop.

Line 16: The second loop of the program is in charge of printing the numbers in reverse order:

for(index=99; index >= 0; index--)

On the first line of the printing loop, the variable index has the value 99. This time around the value of index is decreased by one on each iteration until the value is 0 and the loop has been repeated a hundred times.

Line 17: As the loop is repeated, the values from the array elements are printed on screen, separated with a space. Because the index begins from 99, the number inputted last is the first one to be printed; the numbers are printed in reverse order.